+++ apps/dotlottie-react-example/src/client-entry.tsx
@@ -5,7 +5,9 @@ import App from './App';
 
 ReactDOM.hydrateRoot(
   document.getElementById('root') as HTMLElement,
-  <React.StrictMode>
-    <App />,
-  </React.StrictMode>,
+  // <React.StrictMode>
+  <App />,
+  {
+    /* </React.StrictMode>, */
+  },
 );