+++ packages/web/src/constants.ts
@@ -1,3 +1,5 @@
 export const IS_BROWSER = typeof window !== 'undefined' && typeof window.document !== 'undefined';
 export const MS_TO_SEC_FACTOR = 1000;
 export const DEFAULT_BG_COLOR = '#00000000';
+export const ZIP_SIGNATURE = new Uint8Array([0x50, 0x4b, 0x03, 0x04]);
+export const LOTTIE_JSON_MANDATORY_KEYS = ['v', 'ip', 'op', 'fr', 'w', 'h', 'nm'];