+++ packages/web/src/utils.ts
@@ -1,3 +1,5 @@
+import { LOTTIE_JSON_MANDATORY_KEYS, ZIP_SIGNATURE } from './constants';
+
 export function isHexColor(color: string): boolean {
   return /^#([\da-f]{6}|[\da-f]{8})$/iu.test(color);
 }
@@ -13,4 +15,36 @@ export function hexStringToRGBAInt(colorHex: string): number {
   hex = hex.length === 6 ? `${hex}ff` : hex;
 
   return parseInt(hex, 16);
+}
+
+export function isDotLottie(fileData: ArrayBuffer): boolean {
+  if (fileData.byteLength < 4) {
+    return false;
+  }
+
+  const fileSignature = new Uint8Array(fileData.slice(0, ZIP_SIGNATURE.byteLength));
+
+  for (let i = 0; i < ZIP_SIGNATURE.length; i += 1) {
+    if (ZIP_SIGNATURE[i] !== fileSignature[i]) {
+      return false;
+    }
+  }
+
+  return true;
+}
+
+function isLottieJSON(fileData: Record<string, unknown>): boolean {
+  return LOTTIE_JSON_MANDATORY_KEYS.every((key) => key in fileData);
+}
+
+export function isLottie(fileData: string | Record<string, unknown>): boolean {
+  if (typeof fileData === 'string') {
+    try {
+      return isLottieJSON(JSON.parse(fileData));
+    } catch (_e) {
+      return false;
+    }
+  } else {
+    return isLottieJSON(fileData);
+  }
 }