+++ packages/web/tests/utils.test.ts
@@ -1,6 +1,9 @@
 import { describe, test, expect } from 'vitest';
 
-import { hexStringToRGBAInt, isHexColor } from '../src/utils';
+import { hexStringToRGBAInt, isHexColor, isLottie, isDotLottie } from '../src/utils';
+
+import lottieUrl from './__fixtures__/test.json?url';
+import dotLottieUrl from './__fixtures__/test.lottie?url';
 
 describe('isHexColor', () => {
   test.each([
@@ -39,5 +42,49 @@ describe('hexStringToRGBAInt', () => {
 
   test('handles missing alpha in hex colors', () => {
     expect(hexStringToRGBAInt('#123456')).toBe(0x123456ff);
+  });
+});
+
+describe('isDotLottie', () => {
+  const validZipSignature = new Uint8Array([0x50, 0x4b, 0x03, 0x04]);
+  const invalidZipSignature = new Uint8Array([0x00, 0x00, 0x00, 0x00]);
+
+  test.each([
+    [validZipSignature.buffer, true],
+    [invalidZipSignature.buffer, false],
+    [new ArrayBuffer(2), false],
+  ])('returns %s for %s', (buffer, expected) => {
+    expect(isDotLottie(buffer)).toBe(expected);
+  });
+
+  test('returns true for a valid .lottie URL', async () => {
+    const response = await fetch(dotLottieUrl);
+    const buffer = await response.arrayBuffer();
+
+    expect(isDotLottie(buffer)).toBe(true);
+  });
+});
+
+describe('isLottie', () => {
+  const validLottieJSON = '{"v":"5.5.9","ip":0,"op":300,"fr":30, "w":100, "h":100, "nm":"test"}';
+  const invalidLottieJSON = '{"v":"5.5.9","ip":0,"op":300}, "fr":30, "w":100, "h":100, "nm":"test"}';
+  const validLottieObject = { v: '5.5.9', ip: 0, op: 300, fr: 30, w: 100, h: 100, nm: 'test' };
+  const invalidLottieObject = { v: '5.5.9', ip: 0, op: 300, fr: 30, w: 100, h: 100 };
+
+  test.each([
+    [validLottieJSON, true],
+    [invalidLottieJSON, false],
+    [validLottieObject, true],
+    [invalidLottieObject, false],
+    ['invalid string', false],
+  ])('returns %s for %s', (fileData, expected) => {
+    expect(isLottie(fileData)).toBe(expected);
+  });
+
+  test('returns true for a valid .lottie URL', async () => {
+    const response = await fetch(lottieUrl);
+    const buffer = await response.json();
+
+    expect(isLottie(buffer)).toBe(true);
   });
 });