+++ packages/react/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @lottiefiles/dotlottie-react
 
+## 0.8.1
+
+### Patch Changes
+
+- Updated dependencies [dc66e8e]
+  - @lottiefiles/dotlottie-web@0.29.1
+
+## 0.8.0
+
+### Minor Changes
+
+- eebb307: feat(react): 🎸 add DotLottieWorkerReact component
+
+### Patch Changes
+
+- Updated dependencies [400c333]
+- Updated dependencies [ca7bb5a]
+  - @lottiefiles/dotlottie-web@0.29.0
+
 ## 0.7.2
 
 ### Patch Changes