+++ packages/wc/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @lottiefiles/dotlottie-wc
 
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [dc66e8e]
+  - @lottiefiles/dotlottie-web@0.29.1
+
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies [400c333]
+- Updated dependencies [ca7bb5a]
+  - @lottiefiles/dotlottie-web@0.29.0
+
 ## 0.2.8
 
 ### Patch Changes